requirements.txt
colour
Flask
numpy
Pillow
requests
scipy
tensorflow
Gunicorn
